requirements.txt
flask
sqlalchemy
requests
beautifulsoup
python-memcached
